#8 Test coverage improvements - Factored merge logic into multiple files for improved maintainability - Improve test coverage across multiple components - Renamed many test names to improve usefulness
