+++ src/bundles-processor.test.ts
@@ -4,14 +4,20 @@ import Vinyl from "vinyl";
 import { BundlerStreamFactory } from "./main";
 import { Transform, TransformCallback, Readable } from "stream";
 
-test("BundlesProcessor with iterable inputs empty", async t => {
+/**
+ * Should return empty results.
+ */
+test("Empty inputs", async t => {
     const files: Map<string, [Vinyl, number]> = new Map();
     const bundles: Map<string, string[]> = new Map();
 
     TestBundler(t, [[], new Map()], await BundlesProcessor(files, bundles, BundleStreamFactory, () => {}));
 });
 
-test("BundlesProcessor with files but no bundles", async t => {
+/**
+ * Should return empty results.
+ */
+test("Files but no bundles", async t => {
     const files: Map<string, [Vinyl, number]> = new Map();
     files.set("test", [MakeVinyl("test", "test"), 0]);
     const bundles: Map<string, string[]> = new Map();
@@ -19,7 +25,10 @@ test("BundlesProcessor with files but no bundles", async t => {
     TestBundler(t, [[], new Map()], await BundlesProcessor(files, bundles, BundleStreamFactory, () => {}));
 });
 
-test("BundlesProcessor with files and bundles", async t => {
+/**
+ * Should have tangible results.
+ */
+test("Normal files and bundles", async t => {
     const files: Map<string, [Vinyl, number]> = new Map();
     files.set("test", [MakeVinyl("test", "test"), 0]);
     const bundles: Map<string, string[]> = new Map();
@@ -34,7 +43,10 @@ test("BundlesProcessor with files and bundles", async t => {
     TestBundler(t, [resultChunks, resultPaths], await BundlesProcessor(files, bundles, BundleStreamFactory, () => {}));
 });
 
-test("BundlesProcessor with bundles that will never be satisfied", async t => {
+/**
+ * Should throw an exception indicating that no file could be resolved.
+ */
+test("Unsatisfiable bundles", async t => {
     const files: Map<string, [Vinyl, number]> = new Map();
     files.set("test2", [MakeVinyl("test", "test"), 0]);
     const bundles: Map<string, string[]> = new Map();
@@ -43,7 +55,11 @@ test("BundlesProcessor with bundles that will never be satisfied", async t => {
     await t.throwsAsync(() => BundlesProcessor(files, bundles, BundleStreamFactory, () => {}), 'No file could be resolved for "test".');
 });
 
-test("BundlesProcessor with non-Vinyl chunks emitted by bundle factory", async t => {
+/**
+ * Should return tangible results and complain that a chunk can't be included in results callback.
+ * TODO Verify that complaint is voiced.
+ */
+test("Non-Vinyl chunks emitted by bundle factory", async t => {
     class TestNonVinylTransform extends TestTransform {
         _transform(chunk: any, encoding: string, callback: TransformCallback): void {
             this.push(chunk);
@@ -71,8 +87,11 @@ test("BundlesProcessor with non-Vinyl chunks emitted by bundle factory", async t
     TestBundler(t, [resultChunks, resultPaths], await BundlesProcessor(files, bundles, NonVinylBundleStreamFactory, () => {}));
 });
 
-// TODO Handle exception block in bundles process - promise (if not already)
-test("BundlesProcessor with bundle factory that generates errors (before piping)", async t => {
+/**
+ * Should rethrow the exception generated by the bundle factory.
+ * TODO Verify that errors emitted at the end of bundling are rethrown.
+ */
+test("Bundle factory that throws exception (before stream)", async t => {
     const ErrorBundleStreamFactory: BundlerStreamFactory = (src: Readable): Transform => {
         throw new Error("RIP");
     };
@@ -85,6 +104,12 @@ test("BundlesProcessor with bundle factory that generates errors (before piping)
     await t.throwsAsync(() => BundlesProcessor(files, bundles, ErrorBundleStreamFactory, () => {}), "RIP");
 });
 
+/**
+ * Simplfiies verification of results.
+ * @param t Execution context to run assertions with.
+ * @param expected The expected result.
+ * @param actual The actual result.
+ */
 function TestBundler(t: ExecutionContext, expected: [any[], Map<string, Vinyl[]>], actual: [any[], Map<string, Vinyl[]>]): void {
     // Check result chunks (order insensitive)
     t.deepEqual(expected[0].sort(), actual[0].sort());
@@ -98,7 +123,7 @@ function TestBundler(t: ExecutionContext, expected: [any[], Map<string, Vinyl[]>
 }
 
 /**
- * Simple stub for testing purposes
+ * Simple stub transform stream for testing purposes
  */
 class TestTransform extends Transform {
     constructor() {
@@ -124,6 +149,10 @@ function MakeVinyl(contents: string, path: string): Vinyl {
     });
 }
 
+/**
+ * Bundle factory for testing purposes. Uses test transform.
+ * @param src Source stream.
+ */
 const BundleStreamFactory: BundlerStreamFactory = (src: Readable): Transform => {
     return src.pipe(new TestTransform());
 };
